Interfaz modificada para trabajar solo con enteros. Funcion para sumar en clase Calculadora terminada

+++ src/interfaces/Operaciones.java
@@ -2,6 +2,6 @@ package interfaces;
 
 public interface Operaciones {
 
-    public double hacerOperacion(double num1, double num2);
+    public Integer hacerOperacion(Integer num1, Integer num2);
     
 }
